backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
alembic
python-dotenv
databases
pydantic-settings
